bugfix: clamp_i32 returns an i32 instead of u16

+++ AutonomousDistroBoard/Core/Inc/util.h
@@ -7,7 +7,7 @@
 
 bool util_has_elapsed(uint32_t now, uint32_t start, uint32_t duration_ms);
 uint16_t clamp_u16(uint16_t value, uint16_t min, uint16_t max);
-uint16_t clamp_i32(int32_t value, int32_t min, int32_t max);
+int32_t clamp_i32(int32_t value, int32_t min, int32_t max);
 uint16_t map_u16(uint16_t value, uint16_t in_min, uint16_t in_max, uint16_t out_min, uint16_t out_max);
 int32_t map_i32(int32_t value, int32_t in_min, int32_t in_max, int32_t out_min, int32_t out_max);
